refactor: badges
- BREAKING CHANGE: remove `Blink`, as it's not used
- BREAKING CHANGE: renamed `position` to `tooltipPosition` in `Dot`
- moved each badge to its own directory
- now `stories.tsx` is a correct way to create stories too
- update docs
