+++ .storybook/main.js
@@ -2,6 +2,7 @@ module.exports = {
   "stories": [
     "../src/**/*/stories.@(ts|tsx)",
     "../src/**/*.stories.@(js|jsx|ts|tsx)",
+    "../src/**/stories/index.@(js|jsx|ts|tsx)",
   ],
   "addons": [
     "@storybook/addon-links",
