+++ src/components/badge/index.ts
@@ -1,0 +1,2 @@
+export { Badge } from './Badge';
+export type { Props as BadgeProps } from './Badge';
