+++ src/components/badge/types.ts
@@ -1,0 +1,8 @@
+export type BadgeVariant =
+  | 'light'
+  | 'default'
+  | 'label'
+  | 'success'
+  | 'warning'
+  | 'notice'
+  | 'fail';
