+++ src/components/badge/utils/index.ts
@@ -1,0 +1,2 @@
+export { getBorderColor } from './get-border-color';
+export { getBorderAnimation } from './get-border-animation';
