--- src/components/badges/blink/index.ts
@@ -1,4 +1,0 @@
-import Blink, { BlinkProps } from './blink';
-
-export type { BlinkProps };
-export default Blink;
