--- src/components/badges/index.ts
@@ -1,9 +1,0 @@
-import Badge from './badge';
-import Blink, { BlinkProps } from './blink';
-import Dot from './dot';
-import ProjectBadge from './project-badge';
-import TextValueBadge from './text-value-badge';
-
-export { Badge, Blink, Dot, ProjectBadge, TextValueBadge };
-export type { BadgeProps } from './badge';
-export type { BlinkProps };
