+++ src/components/dot/index.ts
@@ -1,0 +1,3 @@
+export { Dot } from './Dot';
+
+export type { Props as DotProps } from './Dot';
