+++ src/components/project-badge/index.ts
@@ -1,0 +1,3 @@
+export { ProjectBadge } from './ProjectBadge';
+
+export type { Props as ProjectBadgeProps } from './ProjectBadge';
