+++ src/components/text-value-badge/index.ts
@@ -1,0 +1,3 @@
+export { TextValueBadge } from './TextValueBadge';
+
+export type { Props as TextValueBadgeProps } from './TextValueBadge';
