+++ src/components/text-value-badge/styles.ts
@@ -1,0 +1,17 @@
+export const wrapper = {
+  display: 'flex',
+  alignItems: 'center',
+
+  borderWidth: '1px',
+  borderStyle: 'solid',
+
+  borderRadius: '2px',
+
+  overflow: 'hidden',
+
+  '> span': {
+    lineHeight: '13px',
+    py: '3px',
+    px: '6px',
+  },
+};
